하트 획득 로직 추가 및 관련 API 수정 feat. redis

See merge request s08-final/S08P31B307!83

+++ backend/src/main/java/com/chillin/hearting/api/data/HeartConditionData.java
@@ -14,8 +14,8 @@ public class HeartConditionData implements Data {
     private Long heartId;
     private String name;
     private String heartUrl;
-    private Integer currentValue;
-    private Integer maxValue;
+    private Long currentValue;
+    private Long maxValue;
 
     public static HeartConditionData of(Heart heart) {
         return HeartConditionData.builder()
